+++ data/src/main/java/com/kusitms/data/remote/di/RepositoryModule.kt
@@ -1,7 +1,9 @@
 package com.kusitms.data.remote.di
 
-import com.kusitms.data.LoginRepositoryImpl
+import com.kusitms.data.repository.LoginRepositoryImpl
+import com.kusitms.data.repository.SignInRepositoryImpl
 import com.kusitms.domain.repository.LoginRepository
+import com.kusitms.domain.repository.SignInRepository
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -13,4 +15,8 @@ object RepositoryModule {
     @Provides
     fun bindLoginRepository(loginRepositoryImpl: LoginRepositoryImpl):
             LoginRepository = loginRepositoryImpl
+
+    @Provides
+    fun bindSigninRepository(signInRepositoryImpl: SignInRepositoryImpl):
+            SignInRepository = signInRepositoryImpl
 }
