+++ domain/src/main/java/com/kusitms/domain/repository/LoginRepository.kt
@@ -1,10 +1,8 @@
 package com.kusitms.domain.repository
 
-import com.kusitms.domain.model.login.LoginMemberProfile
 
 
 interface LoginRepository {
     suspend fun LoginMember(email: String, password: String): Result<Unit>
 
-    suspend fun fetchLoginMemberProfile(): Result<LoginMemberProfile>
 }
