+++ domain/src/main/java/com/kusitms/domain/repository/SignInRepository.kt
@@ -1,0 +1,7 @@
+package com.kusitms.domain.repository
+
+import com.kusitms.domain.model.login.LoginMemberProfile
+
+interface SignInRepository {
+    suspend fun fetchLoginMemberProfile(): Result<LoginMemberProfile>
+}
